Refactor: extract answergrid as a reusable component
